+++ .prettierrc.js
@@ -1,7 +1,11 @@
 module.exports = {
   arrowParens: 'avoid',
   bracketSameLine: true,
-  bracketSpacing: false,
+  bracketSpacing: true,
+  printWidth: 100,
+  quoteProps: 'consistent',
+  semi: false,
   singleQuote: true,
+  tabWidth: 2,
   trailingComma: 'all',
 };
