+++ src/components/empty-data/ASEmptyData.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View, Text, Image} from 'react-native';
 
-import {ASEmptyDataStyles} from './ASEmptyData-styles';
+import {ASEmptyDataStyles} from './asEmptyData-styles';
 
 const ASEmptyData = () => {
   return (
